Merge branch 'enh-#406-zs' of https://github.com/rossvideo/Catena into enh-#407-zs

+++ sdks/cpp/connections/gRPC/include/BasicParamInfoRequest.h
@@ -81,7 +81,7 @@ class CatenaServiceImpl::BasicParamInfoRequest : public CallData {
          * @param base_path - The base path of the array.
          * @return The length of the array.
          */
-        size_t calculateArrayLength(const std::string& base_path);
+        uint32_t calculateArrayLength(const std::string& base_path);
 
         /**
          * @brief Updates the array lengths of the responses.
@@ -89,7 +89,7 @@ class CatenaServiceImpl::BasicParamInfoRequest : public CallData {
          * @param array_name - The name of the array.
          * @param length - The length of the array.
          */
-        void updateArrayLengths(const std::string& array_name, size_t length);
+        void updateArrayLengths(const std::string& array_name, uint32_t length);
 
 
         /**
@@ -160,10 +160,10 @@ class CatenaServiceImpl::BasicParamInfoRequest : public CallData {
         /**
          * @brief The current response index.
          */
-        size_t current_response_{0};
+        uint32_t current_response_{0};
 
         /**
          * @brief The maximum index of the parameter.
          */
-        size_t max_index_{0};
+        uint32_t max_index_{0};
 };
